chore: update module dependencies
Fixes #2316

Simply update dependencies we don't track on version level to be
compatible with Talos components (like etcd or k8s).

+++ internal/integration/cli/gen.go
@@ -38,8 +38,13 @@ func (suite *GenSuite) SetupTest() {
 }
 
 func (suite *GenSuite) TearDownTest() {
-	suite.Require().NoError(os.Chdir(suite.savedCwd))
-	suite.Require().NoError(os.RemoveAll(suite.tmpDir))
+	if suite.savedCwd != "" {
+		suite.Require().NoError(os.Chdir(suite.savedCwd))
+	}
+
+	if suite.tmpDir != "" {
+		suite.Require().NoError(os.RemoveAll(suite.tmpDir))
+	}
 }
 
 // TestCA ...
